voice_agent/requirements.txt
# Pipecat AI Voice Agent Dependencies
pipecat-ai[google,local,cartesia,silero]>=0.0.90
python-dotenv>=1.0.0
uvicorn>=0.27.0
fastapi>=0.109.0
websockets>=12.0
google-generativeai>=0.3.0
httpx>=0.26.0
pydantic>=2.5.0
